Added support for importing v2 flow events from json
Previously we imported only flow events of version 1 from json.
This CL is adding a support for version 2 flow events, based on
FlowTracker data structure that was added in previous CL.

Bug: 162311144

+++ src/trace_processor/importers/common/flow_tracker.h
@@ -25,7 +25,7 @@
 namespace perfetto {
 namespace trace_processor {
 
-using FlowId = uint32_t;
+using FlowId = uint64_t;
 
 class FlowTracker {
  public:
